frost/requirements.txt
asyncio
discord.py-self
requests
printy
pyfiglet
faker
psutil
discum
colorama
uwuify
brainfuckery
zalgo-text